move forgot password logic to processEmailPasswordReset
